tools/audit/v4/requirements.txt
pyyaml
markdown-it-py
beautifulsoup4
openapi-spec-validator
jsonschema
tabulate
